group-2/project-2/backend/requirements.txt
flask>=3.0
flask-cors>=4.0
openai>=1.0
python-dotenv>=1.0
uuid6>=2024.1
akshare>=1.14
